+++ src/main/java/org/project/soar/model/tag/controller/ChatGPTController.java
@@ -1,9 +1,10 @@
 package org.project.soar.model.tag.controller;
 
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.project.soar.global.api.ApiResponse;
-import org.project.soar.model.tag.YouthPolicyTag;
+import org.project.soar.model.youthpolicytag.YouthPolicyTag;
 import org.project.soar.model.tag.dto.ChatCompletion;
 import org.project.soar.model.tag.service.ChatGPTService;
 import org.springframework.http.HttpStatus;
@@ -34,6 +35,7 @@ public class ChatGPTController {
     chatGPT 유효한 모델인지 조회
      **/
     @GetMapping("/isValidModel")
+    @Operation(summary = "chatGPT 유효한 모델인지 조회")
     public ResponseEntity<ApiResponse<Map<String, Object>>> isValidModel(@RequestParam String modelName) {
         Map<String, Object> result = chatGPTService.isValidModel(modelName);
         ApiResponse<Map<String, Object>> response = ApiResponse.createSuccess(result);
@@ -43,14 +45,16 @@ public class ChatGPTController {
     chatGPT 프롬프트 입력(일반 챗봇)
      **/
     @PostMapping("/prompt")
-    public ResponseEntity<Map<String, Object>> selectPrompt(@RequestBody ChatCompletion chatCompletion) {
+    @Operation(summary = "chatGPT 프롬프트 입력(일반 챗봇)")
+    public ResponseEntity<ApiResponse<Map<String, Object>>> selectPrompt(@RequestBody ChatCompletion chatCompletion) {
         Map<String, Object> result = chatGPTService.prompt(chatCompletion);
-        return ResponseEntity.ok(result);
+        return ResponseEntity.ok(ApiResponse.createSuccess(result));
     }
     /**
     chatGPT Prompt Management(프롬프트 튜닝 태깅)
      **/
     @PostMapping("/promptManagement")
+    @Operation(summary = "youthPolicy_Tag 생성", description = "자동 태깅 실행")
     public ResponseEntity<ApiResponse<List<YouthPolicyTag>>> runPrompt() {
         List<YouthPolicyTag> result = chatGPTService.runPrompt();
         if (result.isEmpty()) {
