+++ src/main/java/org/project/soar/model/tag/service/ChatGPTServiceImpl.java
@@ -7,15 +7,16 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.project.soar.config.RestTemplateConfig;
-import org.project.soar.model.tag.YouthPolicyTag;
+import org.project.soar.model.youthpolicytag.YouthPolicyTag;
 import org.project.soar.model.tag.dto.ChatCompletion;
 import org.project.soar.model.tag.dto.PromptRequest;
-import org.project.soar.model.tag.dto.YouthPolicyTagResponse;
+import org.project.soar.model.youthpolicytag.dto.YouthPolicyTagResponse;
 import org.project.soar.model.tag.repository.TagRepository;
-import org.project.soar.model.tag.repository.YouthPolicyTagRepository;
+import org.project.soar.model.youthpolicytag.repository.YouthPolicyTagRepository;
 import org.project.soar.model.youthpolicy.YouthPolicy;
 import org.project.soar.model.youthpolicy.dto.YouthPolicyOpenAI;
 import org.project.soar.model.youthpolicy.repository.YouthPolicyRepository;
+import org.project.soar.model.youthpolicytag.service.YouthPolicyTagService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
